# docs/swagger.yaml
basePath: /api/v1
definitions:
  controllers.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  controllers.SignInData:
    properties:
      email:
        example: example@gmail.com
        type: string
      password:
        example: "12345678"
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  controllers.SignInResponse:
    properties:
      accessToken:
        type: string
      address:
        type: string
      avatar:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
  models.CreateDishInput:
    properties:
      description:
        example: Phở bò Việt Nam
        type: string
      dishGroupId:
        example: "1"
        type: string
      imageIds:
        example:
        - 7c5a4b8f-fcf6-48d3-b21a-d81ebdfdf6f1
        - 1566e532-72d4-49d4-8fca-c9142816006a
        items:
          type: string
        type: array
      name:
        example: Phở
        type: string
      price:
        example: 50000
        type: number
      status:
        example: 1
        type: integer
    required:
    - dishGroupId
    - imageIds
    - name
    - price
    type: object
...
  models.CreateRestaurantInput:
    properties:
      address:
        example: 36 Pasteur, Ben Nghe, Quan 1, Ho Chi Minh City
        type: string
      email:
        example: example@gmail.com
        type: string
      name:
        example: Example Restaurant
        type: string
      password:
        example: "12345678"
        minLength: 8
        type: string
      phone:
        example: "1234567890"
        type: string
    required:
    - address
    - email
    - name
    - password
    - phone
    type: object
  models.Dish:
    properties:
      createdAt:
        type: string
      description:
        description: optional
        type: string
      dishGroupId:
        type: integer
      id:
        type: integer
      imageIds:
        items:
          type: string
        type: array
      name:
        type: string
      price:
        type: number
      status:
        description: '0: inactive, 1: active'
        type: integer
      updatedAt:
        type: string
    required:
    - dishGroupId
    - name
    - price
    type: object
  models.DishGroup:
    properties:
      createdAt:
        type: string
      dishes:
        items:
          $ref: '#/definitions/models.Dish'
        type: array
      id:
        type: integer
      name:
        type: string
      restaurantId:
        type: string
      updatedAt:
        type: string
    required:
    - name
    type: object
  models.DishGroupInput:
    properties:
      name:
        example: Noodles
        type: string
    required:
    - name
    type: object
  models.DishGroupResponse:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      restaurantId:
        type: string
      updatedAt:
        type: string
    type: object
...
  models.RestaurantResponse:
    properties:
      address:
        type: string
      avatar:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
...
  models.UpdateDishInput:
    properties:
      description:
        example: Phở bò Việt Nam
        type: string
      imageIds:
        example:
        - 7c5a4b8f-fcf6-48d3-b21a-d81ebdfdf6f1
        - 1566e532-72d4-49d4-8fca-c9142816006a
        items:
          type: string
        type: array
      name:
        example: Phở
        type: string
      price:
        example: 50000
        type: number
      status:
        example: 1
        type: integer
    required:
    - imageIds
    - name
    - price
    type: object
...
  models.UpdateRestaurantInput:
    properties:
      address:
        type: string
      avatar:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
      phone:
        type: string
    type: object
info:
  contact: {}
  description: Smart Serve API
  title: Smart Serve
  version: "1.0"
paths:
...
  /auth/me:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.RestaurantResponse'
              type: object
      security:
      - BearerAuth: []
      tags:
      - Auth
  /auth/sign-in:
    post:
      consumes:
      - application/json
      parameters:
      - description: Sign in data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/controllers.SignInData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.Response'
            - properties:
                data:
                  $ref: '#/definitions/controllers.SignInResponse'
              type: object
      tags:
      - Auth
...
  /dish-groups:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.DishGroup'
                  type: array
              type: object
      security:
      - BearerAuth: []
      tags:
      - DishGroups
    post:
      consumes:
      - application/json
      parameters:
      - description: DishGroup Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.DishGroupInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/controllers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.DishGroupResponse'
              type: object
      security:
      - BearerAuth: []
      tags:
      - DishGroups
  /dish-groups/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: DishGroup ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - BearerAuth: []
      tags:
      - DishGroups
    get:
      consumes:
      - application/json
      parameters:
      - description: DishGroup ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.DishGroup'
              type: object
      security:
      - BearerAuth: []
      tags:
      - DishGroups
    put:
      consumes:
      - application/json
      parameters:
      - description: DishGroup ID
        in: path
        name: id
        required: true
        type: string
      - description: DishGroup Data
        in: body
        name: dishGroup
        required: true
        schema:
          $ref: '#/definitions/models.DishGroupInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.DishGroupResponse'
              type: object
      security:
      - BearerAuth: []
      tags:
      - DishGroups
  /dishes:
    get:
      consumes:
      - application/json
      parameters:
      - description: Dish Group ID
        in: query
        name: dishGroupId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Dish'
                  type: array
              type: object
      security:
      - BearerAuth: []
      tags:
      - Dishes
    post:
      consumes:
      - application/json
      parameters:
      - description: Dish Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.CreateDishInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/controllers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Dish'
              type: object
      security:
      - BearerAuth: []
      tags:
      - Dishes
  /dishes/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Dish ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.Response'
            - properties:
                data:
                  type: object
              type: object
      security:
      - BearerAuth: []
      tags:
      - Dishes
    get:
      consumes:
      - application/json
      parameters:
      - description: Dish ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Dish'
              type: object
      security:
      - BearerAuth: []
      tags:
      - Dishes
    put:
      consumes:
      - application/json
      parameters:
      - description: Dish ID
        in: path
        name: id
        required: true
        type: string
      - description: Dish Data
        in: body
        name: DishInput
        required: true
        schema:
          $ref: '#/definitions/models.UpdateDishInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Dish'
              type: object
      security:
      - BearerAuth: []
      tags:
      - Dishes
...
  /restaurants:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Restaurant Data
        in: body
        name: restaurant
        required: true
        schema:
          $ref: '#/definitions/models.UpdateRestaurantInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controllers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.RestaurantResponse'
              type: object
      security:
      - BearerAuth: []
      tags:
      - Restaurants
    post:
      consumes:
      - application/json
      parameters:
      - description: Restaurant Data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.CreateRestaurantInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/controllers.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.RestaurantResponse'
              type: object
      tags:
      - Restaurants
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"